+++ pyproject.toml
@@ -14,7 +14,7 @@ log_file_format = "%(asctime)s.%(msecs)03d %(levelname)-8s %(message)s"
 log_file_date_format = "%H:%M:%S"
 
 junit_family = "xunit2"
-junit_logging = "system-err"
+junit_logging = "no"
 junit_duration_report = "call"
 junit_suite_name = "system_tests_suite"
